pyproject.toml
[tool.poetry]
name = "parser-codeforces"
version = "0.1.0"
description = ""
authors = ["violetta <lotsmanovavioletta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pytelegrambotapi = "^4.14.0"
schedule = "^1.2.1"
flake8 = "^6.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
